Replace int256 with int

+++ packages/loopring_v3/contracts/iface/IBlockVerifier.sol
@@ -50,11 +50,11 @@ contract IBlockVerifier
     /// @param blockVersion The block version (i.e. which circuit version needs to be used)
     /// @param vk The verification key
     function registerCircuit(
-        uint8  blockType,
-        bool   onchainDataAvailability,
-        uint16 blockSize,
-        uint8  blockVersion,
-        uint256[18] calldata vk
+        uint8    blockType,
+        bool     onchainDataAvailability,
+        uint16   blockSize,
+        uint8    blockVersion,
+        uint[18] calldata vk
         )
         external;
 
@@ -86,12 +86,12 @@ contract IBlockVerifier
     /// @param proofs The ZK proofs proving that the blocks are correct
     /// @return True if the block is valid, false otherwise
     function verifyProofs(
-        uint8   blockType,
-        bool    onchainDataAvailability,
-        uint16  blockSize,
-        uint8   blockVersion,
-        uint256[] calldata publicInputs,
-        uint256[] calldata proofs
+        uint8  blockType,
+        bool   onchainDataAvailability,
+        uint16 blockSize,
+        uint8  blockVersion,
+        uint[] calldata publicInputs,
+        uint[] calldata proofs
         )
         external
         view
