+++ packages/loopring_v3/contracts/test/TransferContract.sol
@@ -37,7 +37,7 @@ contract TransferContract {
     function safeTransferWithGasLimit(
         address token,
         address to,
-        uint256 value,
+        uint value,
         uint    gasLimit
         )
         external
@@ -52,7 +52,7 @@ contract TransferContract {
         address token,
         address from,
         address to,
-        uint256 value,
+        uint value,
         uint    gasLimit
         )
         external
